web_test/requirements.txt
flask>=2.3.0
langchain-openai
langchain-core
python-dotenv
asyncio